+++ app/layout.tsx
@@ -1,6 +1,5 @@
-import { Clarity } from "@/components/Clarity";
+import { Clarity } from "@/components/clarity-script";
 import { Toaster } from "@/components/ui/sonner";
-
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import Link from "next/link";
